Fix name of test in junit not present when setup fails
This change is better understood with an example. Let's assume the following
test case:

setup_file() {
  false
}

@test "This is a Test" {
  bats_require_minimum_version 1.5.0
  true
}

Running this test suite with the junit report would give us the following

$ ./bin/bats  --print-output-on-failure --formatter junit example.bats
<?xml version="1.0" encoding="UTF-8"?>
<testsuites time="0">
<testsuite name="example.bats" tests="1" failures="1" errors="0" skipped="0" time="0" timestamp="2025-07-28T19:02:38" hostname="tsumugi.local">
    <testcase classname="example.bats" name="" time="0">
        <failure type="failure">(from function `setup_file&#39; in test file example.bats, line 4)
  `false&#39; failed</failure>
    </testcase>

</testsuite>
</testsuites>

Note that the "name" field from the test is empty. This is an issue
because fundamentally, if the setup failed, the test itself has failed.
Also, most junit parsers (gitlab, x-ray) expect a name regardless of
when the failure has occurred.

We can use the TAP reporter, which is the "fundamental" reporter of BATS
to check how this is currently being parsed by the junit reporter

$ ./bin/bats  --print-output-on-failure --formatter tap example.bats
1..1
not ok 1 setup_file failed

Which provides no information about which test we were trying to run when
the setup failed, even though that reporting is on the test-level, not on
the suite level.

After this change, we exploit the comment field in the TAP standard as
follows, and actually return the test name itself

$ ./bin/bats  --print-output-on-failure --formatter tap example.bats
1..1
not ok 1 test_This_is_a_Test # setup_file failed

Which can be parsed by the junit exporter, generating the now correct junit
format

$ ./bin/bats  --print-output-on-failure --formatter junit example.bats
<?xml version="1.0" encoding="UTF-8"?>
<testsuites time="0">
<testsuite name="example.bats" tests="1" failures="1" errors="0" skipped="0" time="0" timestamp="2025-07-28T19:08:08" hostname="tsumugi.local">
    <testcase classname="example.bats" name="This is a Test" time="0">
        <failure type="failure">setup_file failed
(from function `setup_file&#39; in test file example.bats, line 4)
  `false&#39; failed</failure>
    </testcase>

</testsuite>
</testsuites>

Adding more tests will not report all tests as failed, just the first one

$ cat example.bats

setup_file() {
  false
}

@test "This is a Test" {
  bats_require_minimum_version 1.5.0
  true
}

@test "This is another Test" {
  bats_require_minimum_version 1.5.0
  true
}

$ ./bin/bats  --print-output-on-failure --formatter tap example.bats
1..2
not ok 1 test_This_is_another_Test # setup_file failed

As it seems more correct to fail the first test fast if the setup failed.

Naturally, this is a trade-off in the design of this behaviour.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>