fix(formatter): skipped tests reporting time as part of the name
This bug was introduced by commit 6314cc4 and does not exist in the upstream
bats-core. The "current_test_name" variable in bats has the following format
for executed tests:

(ok/not ok) My test name in 10 ms

But for skipped tests, the format is:

ok My test name in 10 ms # skip reason

When commit 6314cc4 refactored "formatter.bash", the script began assuming that
the test duration would always appear at the end of the "current_test_name"
string, which is not true for skipped tests. Because of this, the timing ended
up being treated as part of the test name.

In this case, the test name would incorrectly be identified as:

My test name in 10 ms

Related-to: TCCP-1023

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>